+++ apps/web/src/app/[lang]/app/[type]/contracts/refund/refund-fees/[id]/edit.tsx
@@ -31,7 +31,7 @@ import {
   postRefundTableFeeHeaderDetailsByRefundTableHeaderId,
   putRefundTableFeeHeaderDetailsById,
   putRefundTableFeeHeadersById,
-} from "../../../action";
+} from "../../../actions/refund-tables";
 
 export default function Edit({
   details,
